New layer of abstraction: the Environment
The Environment class holds a Board within in. The Environment is
responsible for knowing whose turn it is, "scoring" moves that are made,
and making sure that the board's state representation is player-aware.

I've also added a Player abstract class, with HumanPlayer and a dummy
AIPlayer extending Player.

Finally, everything is run from a Controller class, which manages
initializing players (Human or AI), asking players for moves, passing
the moves down to the Environment (which further pushes the move down to
Board), and managing the graphical interface.